clean up directory
