+++ apps/socials-bot/src/routes/hello-world.ts
@@ -1,0 +1,9 @@
+import { Hono } from "hono";
+
+import { HonoConfig } from "~/config";
+
+export const helloWorld = new Hono<HonoConfig>().get("/", (c) => {
+  return c.json({
+    message: "Hello, world!",
+  });
+});
