+++ apps/socials-bot/src/utils.ts
@@ -1,3 +1,4 @@
 export function isAdmin(roles: string[], adminRoleId: string) {
   return roles.includes(adminRoleId);
 }
+
